bookstore/requirements.txt
PyJWT >= 2.3.0